only redis failing
