+++ app/apollo.js
@@ -1,8 +1,6 @@
 import { ApolloServer } from '@apollo/server'
 //const { RedisCache } = require('apollo-server-cache-redis')
 import { makeExecutableSchema } from '@graphql-tools/schema'
-import cors from 'cors'
-import { ResolverCacheDataSource } from '@digicatapult/resolver-cache-datasource'
 import validationPlugin from '@digicatapult/apollo-type-validation-plugin'
 
 const {
@@ -13,9 +11,6 @@ const {
 import typeDefs from './graphql/typeDefs.js'
 import resolvers from './graphql/resolvers.js'
 import env from './env.js'
-import logger from './logger.js'
-import buildDataLoaders from './loaders/index.js'
-import { users as usersService } from './services/index.js'
 
 function createApolloServer() {
   const schema = makeExecutableSchema({
@@ -26,7 +21,7 @@ function createApolloServer() {
 
   const server = new ApolloServer({
     schema,
-    cors: cors(),
+    validationRules: [],
     playground: env.ENABLE_GRAPHQL_PLAYGROUND,
     tracing: env.ENABLE_GRAPHQL_PLAYGROUND,
     plugins: [typeValidationPlugin({ schema, directives: [arrayLengthDirective(), boundedIntegerDirective()] })],
@@ -37,26 +32,6 @@ function createApolloServer() {
     //   ...(env.CACHE_PASSWORD !== null ? { password: env.CACHE_PASSWORD } : {}),
     //   ...(env.CACHE_ENABLE_TLS ? { tls: {} } : {}),
     // }),
-    dataSources: () => {
-      return {
-        autoResolver: new ResolverCacheDataSource({
-          defaultTTL: env.CACHE_MAX_TTL,
-        }),
-      }
-    },
-    context: async ({ req }) => {
-      let user = null
-      try {
-        const userId = req.headers['user-id']
-        user = await usersService.getCurrentUser(userId)
-      } catch (err) {
-        user = null
-      }
-
-      const loaders = buildDataLoaders()
-
-      return { logger, user, loaders }
-    },
   })
 
   return server
