+++ app/server.js
@@ -3,7 +3,10 @@ import cors from 'cors'
 import compression from 'compression'
 import pinoHttp from 'pino-http'
 import { expressMiddleware } from '@apollo/server/express4'
+import { ResolverCacheDataSource } from '@digicatapult/resolver-cache-datasource'
 
+import buildDataLoaders from './loaders/index.js'
+import { users as usersService } from './services/index.js'
 import env from './env.js'
 import logger from './logger.js'
 import createApolloServer from './apollo.js'
@@ -14,7 +17,6 @@ async function createHttpServer() {
   const app = express()
 
   app.use(compression())
-  app.use(cors())
 
   app.use((req, res, next) => {
     if (req.path !== '/health') requestLogger(req, res)
@@ -22,17 +24,37 @@ async function createHttpServer() {
   })
 
   await server.start()
-  app.use(
-    cors(),
-    express.json({ limit: env.MAX_QUERY_SIZE }),
-    expressMiddleware(server, {
-      context: async ({ req }) => ({ token: req.headers.token }),
-    })
-  )
 
   app.get('/health', async (req, res) => {
     res.status(200).send({ status: 'ok' })
   })
+
+  app.use(
+    cors(),
+    express.json({ limit: env.MAX_QUERY_SIZE }),
+    expressMiddleware(server, {
+      context: async ({ req }) => {
+        let user = null
+        try {
+          const userId = req.headers['user-id']
+          user = await usersService.getCurrentUser(userId)
+        } catch (err) {
+          user = null
+        }
+
+        const loaders = buildDataLoaders()
+
+        return { logger, user, loaders }
+      },
+      dataSources: () => {
+        return {
+          autoResolver: new ResolverCacheDataSource({
+            defaultTTL: env.CACHE_MAX_TTL,
+          }),
+        }
+      },
+    })
+  )
 
   app.use((err, req, res, next) => {
     if (!res.headersSent) {
